AAAAAA

+++ src/main/java/com/ali_ahmad_fahrezy_187221042/alproprak/Week9/Soal1_187221042_Week9_1.java
@@ -25,13 +25,13 @@ public class Soal1_187221042_Week9_1 {
         double x = input.nextDouble();
 
         //Ibu mendapat 1/4 bagian
-        double ibu = (double) 1/4;
+        double ibu = 1d/4;
 
         //Sisanya adalah 3/4
         double sisa = 1 - ibu;
 
         //Satu anak mendapatkan 1/3 dari sisa tersebut
-        double anak = (double) 1/3*sisa;
+        double anak = 1d/3*sisa;
 
         //Kalkulasi warisan
         double warisan = x/(ibu + anak);
